% Generated by build script
\name{chi_metadata}
\alias{chi_metadata}
\title{Get metadata for a specific dataset on Chicago Open Data}
\usage{
chi_metadata()
}
\arguments{
\item{dataset_id}{Socrata dataset identifier (e.g. "ijzp-q8t2")}
}
\value{tibble: id, name, description, category, type, updated_at,}
\description{Get metadata for a specific dataset on Chicago Open Data}
